Added `useTotalPoolsBalance` hook
