+++ src/hooks/usePoolBalance.ts
@@ -3,7 +3,7 @@ import { useQuery, useQueries } from "react-query";
 import { Pool } from "../utils/poolUtils";
 import { useRari } from "../context/RariContext";
 import Rari from "../rari-sdk/index";
-import { BN, stringUsdFormatter } from "../utils/bigUtils";
+import { BN } from "../utils/bigUtils";
 import { getSDKPool } from "../utils/poolUtils";
 
 export const fetchPoolBalance = async ({
@@ -14,14 +14,14 @@ export const fetchPoolBalance = async ({
   pool: Pool;
   rari: Rari;
   address: string;
-}) : Promise<BN> => {
+}): Promise<BN> => {
   const balance = await getSDKPool({ rari, pool }).balances.balanceOf(address);
 
-  let formattedBalance = stringUsdFormatter(rari.web3.utils.fromWei(balance));
+  // let formattedBalance = stringUsdFormatter(rari.web3.utils.fromWei(balance));
 
-  if (pool === Pool.ETH) {
-    formattedBalance = formattedBalance.replace("$", "") + " ETH";
-  }
+  // if (pool === Pool.ETH) {
+  //   formattedBalance = formattedBalance.replace("$", "") + " ETH";
+  // }
 
   return balance;
 };
@@ -60,4 +60,31 @@ export const usePoolBalances = (pools: Pool[]) => {
           ({ isLoading, error, data })
       )
     , [poolBalances])
+}
+
+export const useTotalPoolsBalance = () => {
+  const { rari, address, } = useRari();
+
+  const { isLoading, data, error } = useQuery(
+    address + " allPoolBalance",
+    async () => {
+      const [stableBal, yieldBal, ethBalInETH, ethPriceBN] = await Promise.all([
+        rari.pools.stable.balances.balanceOf(address),
+        rari.pools.yield.balances.balanceOf(address),
+        rari.pools.ethereum.balances.balanceOf(address),
+        rari.getEthUsdPriceBN(),
+      ]);
+
+      const ethBal = ethBalInETH.mul(
+        ethPriceBN.div(rari.web3.utils.toBN(1e18))
+      );
+
+      return parseFloat(
+        rari.web3.utils.fromWei(stableBal.add(yieldBal).add(ethBal))
+      );
+    }
+  );
+
+  return { isLoading, data, error}
+
 }
